+++ lib/widgets/dialogs/entry_editors/edit_date_dialog.dart
@@ -13,6 +13,7 @@ import 'package:aves/utils/time_utils.dart';
 import 'package:aves/widgets/common/basic/text_dropdown_button.dart';
 import 'package:aves/widgets/common/basic/wheel.dart';
 import 'package:aves/widgets/common/extensions/build_context.dart';
+import 'package:aves/widgets/common/fx/transitions.dart';
 import 'package:aves/widgets/common/providers/media_query_data_provider.dart';
 import 'package:aves/widgets/dialogs/aves_dialog.dart';
 import 'package:aves/widgets/dialogs/item_picker.dart';
@@ -111,7 +112,7 @@ class _EditEntryDateDialogState extends State<EditEntryDateDialog> {
                 duration: context.read<DurationsData>().formTransition,
                 switchInCurve: Curves.easeInOutCubic,
                 switchOutCurve: Curves.easeInOutCubic,
-                transitionBuilder: _formTransitionBuilder,
+                transitionBuilder: AvesTransitions.formTransitionBuilder,
                 child: Column(
                   key: ValueKey(_action),
                   mainAxisSize: MainAxisSize.min,
@@ -142,15 +143,6 @@ class _EditEntryDateDialogState extends State<EditEntryDateDialog> {
       ),
     );
   }
-
-  Widget _formTransitionBuilder(Widget child, Animation<double> animation) => FadeTransition(
-        opacity: animation,
-        child: SizeTransition(
-          sizeFactor: animation,
-          axisAlignment: -1,
-          child: child,
-        ),
-      );
 
   Widget _buildSetCustomContent(BuildContext context) {
     final l10n = context.l10n;